+++ apps/dashboard/app/(components)/welcome-card/index.tsx
